move methods to ns application
